+++ src/hooks/index.ts
@@ -1,6 +1,7 @@
 export * from "./useRide";
 export * from "./useRides";
 export * from "./useJoinRide";
+export * from "./useAddAnonymousUser";
 export * from "./useAddRide";
 export * from "./useAddPacelineRides";
 export * from "./useEditRide";