+++ src/hooks/useAddAnonymousUser.ts
@@ -1,0 +1,45 @@
+import useSWR from "swr";
+import { AnonymousUserValues } from "../components/forms/AnonymousUserForm";
+
+export const addAnonymousUser = async (userAndRide: AnonymousUserValues) => {
+  const { rideId, ...user } = userAndRide;
+
+  try {
+    // Create an anonymous user record
+    const userResult = await fetch(`/api/user/create-anonymous`, {
+      method: "POST",
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(user),
+    }).then((res) => res.json());
+    // Get the userId
+    const { id: userId } = userResult;
+    // Create a UserToRide join record
+    const joinResult = await fetch(`/api/ride/join-anonymous`, {
+      method: "POST",
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ userId, rideId }),
+    }).then((res) => res.json());
+
+    return joinResult;
+  } catch (err) {
+    return { error: err };
+  }
+};
+
+export const useAddAnonymousUser = (user: AnonymousUserValues) => {
+  const { data, error } = useSWR(`/api/user/create-anon`, () =>
+    addAnonymousUser(user)
+  );
+
+  return {
+    user: data,
+    loading: !error && !data,
+    error,
+  };
+};