+++ src/pages/api/ride/join.ts
@@ -47,3 +47,18 @@ const joinRide = async (req: NextApiRequest, res: NextApiResponse) => {
 };
 
 export default joinRide;
+
+export const joinRideAnonymously = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {
+  try {
+    const { rideId, userId } = req.body;
+    const success = await addRiderToRide({ rideId, userId });
+    return res.status(200).json(success);
+  } catch (err) {
+    return res.status(401).send({
+      error: "Not authorised to use this API",
+    });
+  }
+};