+++ src/pages/ride/[id]/join.tsx
@@ -1,0 +1,77 @@
+import type { NextPage } from "next";
+import Head from "next/head";
+import { useRouter } from "next/router";
+import { useState } from "react";
+import { useForm, SubmitHandler } from "react-hook-form";
+import { useSWRConfig } from "swr";
+import { addAnonymousUser } from "../../../hooks";
+import { AnonymousUserForm, AnonymousUserValues } from "../../../components";
+import { flattenQuery } from "../../../../shared/utils";
+
+const JoinRidePage: NextPage = () => {
+  const { mutate } = useSWRConfig();
+  const router = useRouter();
+  const {
+    query: { id: rideId },
+  } = router;
+  const [waiting, setWaiting] = useState(false);
+
+  const {
+    register,
+    handleSubmit,
+    formState: { errors, isDirty },
+  } = useForm<AnonymousUserValues>();
+
+  // Initial state for form: set name, leader and time
+  const defaultValues = {
+    name: "",
+    mobile: "",
+    emergency: "",
+    rideId: "",
+  };
+
+  const onSubmit: SubmitHandler<AnonymousUserValues> = async ({
+    name,
+    mobile,
+    emergency,
+  }) => {
+    setWaiting(true);
+    // Transform data before sending
+    const results = await mutate(`/api/ride/${rideId}`, () =>
+      addAnonymousUser({
+        name,
+        mobile,
+        emergency,
+        rideId: flattenQuery(rideId),
+      })
+    );
+    if (results.userId) {
+      router.back();
+    }
+  };
+
+  return (
+    <>
+      <Head>
+        <title>Join BCC Ride</title>
+        <meta name="description" content="Join Bath Cycling Club Ride" />
+      </Head>
+      <div className="w-full text-neutral-800">
+        <div className="flex w-full flex-row items-center justify-center bg-blue-900 p-2 font-bold uppercase tracking-wide text-white sm:rounded">
+          Join Ride
+        </div>
+
+        <AnonymousUserForm
+          defaultValues={defaultValues}
+          errors={errors}
+          isDirty={isDirty}
+          register={register}
+          handleSubmit={handleSubmit(onSubmit)}
+          waiting={waiting}
+        />
+      </div>
+    </>
+  );
+};
+
+export default JoinRidePage;